requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
yt-dlp>=2024.01.01
supabase>=2.0.0
python-dotenv>=1.0.0
